requirements.txt
pywebview>=5.0
openpyxl>=3.1
